agents/requirements.txt
google-genai
ollama
requests
beautifulsoup4
pymupdf
pandas
streamlit
PyGithub
google-api-core
